Adjusted boxes, alignments, and banner
I adjusted the height of the rounded boxes. I also aligned the dropdown menu in the campaign, the text in the biography box, the dropdown buttons, and the timeline. I also made sure the banner covered the whole width of the website and flushed the rounded boxes. Lastly, I changed the background color of the website from white to a light blue.